+++ TestUtils.ts
@@ -67,7 +67,6 @@ const DEFAULT_GLOBAL_CONFIG: Config.GlobalConfig = {
 
 const DEFAULT_PROJECT_CONFIG: Config.ProjectConfig = {
   automock: false,
-  browser: false,
   cache: false,
   cacheDirectory: '/test_cache_dir/',
   clearMocks: false,
@@ -83,9 +82,7 @@ const DEFAULT_PROJECT_CONFIG: Config.ProjectConfig = {
   globalSetup: null,
   globalTeardown: null,
   globals: {},
-  haste: {
-    providesModuleNodeModules: [],
-  },
+  haste: {},
   moduleDirectories: [],
   moduleFileExtensions: ['js'],
   moduleLoader: '/test_module_loader_path',