+++ packages/jest-config/src/__tests__/readConfigs.test.ts
@@ -7,8 +7,6 @@
 
 import {readConfigs} from '../index';
 
-jest.mock('../importEsm', () => (s: string) => import(s));
-
 test('readConfigs() throws when called without project paths', async () => {
   await expect(
     // @ts-ignore