+++ scripts/verifyOldTs.js
@@ -32,7 +32,7 @@ const cwd = tempy.directory();
 
 try {
   execa.sync('yarn', ['init', '--yes'], {cwd, stdio: 'inherit'});
-  execa.sync('yarn', ['add', 'typescript@~3.4'], {cwd, stdio: 'inherit'});
+  execa.sync('yarn', ['add', 'typescript@~3.8'], {cwd, stdio: 'inherit'});
   fs.writeFileSync(
     path.join(cwd, 'tsconfig.json'),
     JSON.stringify(tsConfig, null, 2),